refactor: fill out working example
This commit adds a working example of a HTTP handler and some docs to
go with it.
