+++ types/wit.d.ts
@@ -1,0 +1,7 @@
+import { WasiClocksMonotonicClock } from './interfaces/wasi-clocks-monotonic-clock.js';
+import { WasiHttpTypes } from './interfaces/wasi-http-types.js';
+import { WasiIoError } from './interfaces/wasi-io-error.js';
+import { WasiIoPoll } from './interfaces/wasi-io-poll.js';
+import { WasiIoStreams } from './interfaces/wasi-io-streams.js';
+import { WasiHttpIncomingHandler } from './interfaces/wasi-http-incoming-handler.js';
+export const incomingHandler: typeof WasiHttpIncomingHandler;
