+++ wit/component.wit
@@ -1,0 +1,5 @@
+package experiments:wasm-lambda;
+
+world component {
+  export wasi:http/incoming-handler@0.2.0;
+}
